a million changes

+++ learning/domains/pushing/verify_data_generation.py
@@ -1,4 +1,4 @@
-import math
+
 import tqdm
 from functools import partial
 import trimesh
@@ -8,27 +8,23 @@ import os
 import pickle
 from types import SimpleNamespace
 import numpy as np
-from scipy.spatial.transform import Rotation as R 
 
+# from learning.domains.pushing.find_contact_points import find_contact_point_and_check_push
 from learning.domains.pushing.find_contact_points import (
     find_contact_point_and_check_push,
-    run_sim,
 )
 from pb_robot.planners.antipodalGraspPlanner import (
     GraspSimulationClient,
     GraspableBody,
     GraspableBodySampler,
-
 )
-
-import pb_robot 
 import pybullet as p
 
 # Don't know if these are correct for now, mostly filler
-PUSH_VELOCITY_RANGE = (0.08, 0.12)
-MASS_RANGE = (0.1, 0.3)
-FRICTION_RANGE = (0.05, 0.1)
-OFFSET_STD_DEV = 0.025
+PUSH_VELOCITY_RANGE = (0.25 / 2.5, 0.3 / 2.5)
+MASS_RANGE = (0.1, 0.2)
+FRICTION_RANGE = (8*0.1, 8*0.3)
+OFFSET_RANGE = (-0.02, 0.02)
 
 
 def get_object_list(fname):
@@ -38,6 +34,11 @@ def get_object_list(fname):
         objects = handle.readlines()
     return [o.strip() for o in objects if len(o) > 1]
 
+
+# From Seiji:
+# Set up data directory.
+# Good practice: save the running arguments (either as a .pkl or .json)
+# so you can remember your settings! A future you will thank you.
 def save_args(new_args):
     """
     Save the arguments for generating the datasets so that it doesn't have to be repeated in the future.
@@ -99,6 +100,10 @@ def generate_parameters(name):
 
     # This function is essentially a black-box for now
     com = GraspableBodySampler._sample_com(name)  # Taking this from GNP
+    # print(name)
+    # print(mass)
+    # print(friction)
+    # print(com)
     return {"name": name, "mass": mass, "friction": friction, "com": com}
 
 
@@ -117,103 +122,23 @@ def generate_object_parameters(args):
         with open(file_name, "wb") as handle:
             # print("HERE")
             pickle.dump(list_of_objects[i], handle)
-    print("Done generating", args.directory_name)
+    print("DONE GENERATING ", args.directory_name)
 
 
 def process_single_object(obj_data, args):
     """
-    A new way of simulating the pushing action that uses the robot. 
-    This is designed to create a simulation that will have more varied results, 
-    as well as being more realistic. 
-    Args: 
-        obj_data: The data for the object that is being simulated.
-        args: The arguments for the simulatio, which is inputted into the file.
-    Returns:
-        data: The data for the object that was simulated.
-    """
-    body = GraspableBody(
-        obj_data["name"], obj_data["com"], obj_data["mass"], obj_data["friction"]
-    )
-
-    sim_client = GraspSimulationClient(body, False)
-    urdf = sim_client._get_object_urdf(body)
-    sim_client.disconnect() 
-
-    data = [] 
-    # TODO: Fix cases where the robot acts weirdly 
-    # TODO: Check if offset is reasonable
-    num_fallen = 0
-    for _ in range(args.n_pushes_per_object):
-        contact_points = None 
-        n_attempts = 0
-        while contact_points is None: 
-            n_attempts += 1
-            if n_attempts > 5:
-                # transformation, contact_points, initial = run_sim(urdf, push_angle, object_angle, push_velocity, offset, gui=True) 
-                break
-            push_angle = np.random.uniform(0, 2 * math.pi) 
-            object_angle = np.random.uniform(0, 2 * math.pi) 
-            offset = np.random.normal(loc=0, scale=OFFSET_STD_DEV)
-            push_velocity = np.random.uniform(*PUSH_VELOCITY_RANGE) 
-            # print("DEBUG MODE")
-            # push_angle = math.pi / 6 
-            object_angle = 0 
-            offset = 0 
-            push_velocity = 0.1 
-
-            # print(push_angle, object_angle, push_velocity, offset)
-            
-            transformation, contact_points, initial, fallen = run_sim(urdf, push_angle, object_angle, push_velocity, offset, gui=args.gui) 
-            if contact_points is not None: 
-                initial_euler_angles = R.from_matrix(initial[:3,:3]).as_euler('xyz', degrees=False)
-                if initial_euler_angles[0] < -0.4 or initial_euler_angles[0] > 0.4 or initial_euler_angles[1] < -0.4 or initial_euler_angles[1] > 0.4:
-                    print("This case")
-                    contact_points = None 
-                    continue
-
-            # print(obj_data, transformation[:3, 3])
-        # print(n_attempts)
-
-        if contact_points is not None:
-            # print(R.from_matrix(initial[:3,:3]).as_euler('xyz', degrees=False), object_angle)
-            data.append(((push_angle, contact_points[0], contact_points[1], body, push_velocity, initial), (transformation, fallen)))
-            num_fallen += fallen 
-        else: 
-            data.append(((None, None, None, None, None, None), (None, None))) 
-            num_fallen += 1
-
-    # if num_fallen == 0:
-    #     print("No fallen objects for ", obj_data)
-    # print(num_fallen, "/", args.n_pushes_per_object)
-    return data 
-
-def process_single_object_old(obj_data, args):
-    """
     Generate the pushing actions for a singular object. This uses a simple pushing action and this function allows
     easy use of multiprocessing for helping the data generation process speed up.
-
-    WARNING: This is old and should no longer really be used.
     """
-    # print("????")
     total_success = 0
     total_fail = 0
     # print(obj_data)
     body = GraspableBody(
         obj_data["name"], obj_data["com"], obj_data["mass"], obj_data["friction"]
     )
-    print(obj_data["com"]) 
-
     sim_client = GraspSimulationClient(body, False)
     urdf = sim_client._get_object_urdf(body)
-    
-
-    # run_sim(urdf) 
-
-    # print(urdf)) 
-
-    # sim_client.disconnect()
-    # run_sim(urdf, np.random.uniform(2*math.pi), np.random.uniform(2*math.pi))
-    # return
+    sim_client.disconnect()
 
     data = []
     for _ in range(args.n_pushes_per_object):
@@ -222,42 +147,31 @@ def process_single_object_old(obj_data, args):
         offset = [np.random.uniform(*OFFSET_RANGE), np.random.uniform(*OFFSET_RANGE), 0]
 
 
-        contact_point = None
-        amount = 0
-        while contact_point is None: 
-            amount += 1 
-            if amount > 10:
-                break
-            angle = np.random.uniform(0, 360)
-            push_velocity = np.random.uniform(*PUSH_VELOCITY_RANGE)
-            offset = [np.random.uniform(*OFFSET_RANGE), np.random.uniform(*OFFSET_RANGE), 0]
-            # print("???")
-            # print(args.gui)
-            print(urdf)
-            contact_point, normals, cube_orn_at_contact, success, logs = find_contact_point_and_check_push(
-                urdf,
-                angle,
-                push_velocity,
-                obj_data["mass"],
-                obj_data["friction"],
-                obj_data["com"],
-                offset,
-                logging=True,
-                gui=args.gui
-            )
-        # if contact_point is None: 
-        #     print("NO CONTACT POINT") 
-        #     print(obj_data) 
-        #     find_contact_point_and_check_push(urdf, angle, push_velocity, obj_data["mass"], obj_data["friction"], obj_data["com"], offset, gui=True)
-        #     print("?")
-
+        for i in range(10):
+            contact_point = None
+            while contact_point is None: 
+                contact_point, normals, success, logs = find_contact_point_and_check_push(
+                    urdf,
+                    angle,
+                    push_velocity,
+                    obj_data["mass"],
+                    obj_data["friction"],
+                    obj_data["com"],
+                    offset,
+                    logging=True,
+                    gui=args.gui
+                )
+            
+        #     print(logs[-1][:3]) 
+        # print("------------------")
+            # print("?")
         # print("DONE")
         # _, _, logs2 = find_contact_point_and_check_push(urdf, angle, push_velocity, obj_data['mass'], obj_data['friction'], obj_data['com'], offset, logging=True)
         # for i in range(len(logs2)):
         #     if logs[i][0] - logs2[i][0] > 0.1:
         #         print("DIFFERENT")
         # move_on = input('Move on!')
-        data.append(((angle, contact_point, normals, body, push_velocity, cube_orn_at_contact), (success, logs)))
+        data.append(((angle, contact_point, normals, body, push_velocity), (success, logs)))
         # exit()
         # if success:
         #     find_contact_point_and_check_push(urdf, angle, push_velocity, obj_data['mass'], obj_data['friction'], obj_data['com'], offset, gui=True)
@@ -270,12 +184,9 @@ def process_single_object_old(obj_data, args):
 
     return data
 
-def init_pool(): 
-    np.random.seed() 
 
 def main(args):
     args = save_args(args)
-    print(args)
     train_objects, test_objects = get_object_lists(args)
     data_root_path = os.path.join("learning", "data", "pushing", args.data_root_name)
     # print(train_objects)
@@ -313,32 +224,32 @@ def main(args):
         )
         generate_object_parameters(test_samegeo_args)
 
-    num_processes = 4 if not args.gui else 1 
-    # num_processes = 1
+    # num_processes = 4 if not args.gui else 1 
+    num_processes = 1
     print("Num Processes: ", num_processes)
-    pool = multiprocessing.Pool(processes=num_processes, initializer=init_pool)
+    pool = multiprocessing.Pool(processes=num_processes)
     process_func = partial(process_single_object, args=args)
     object_data_list = []
 
     print(len(train_objects), len(test_objects))
 
-    if not os.path.exists(os.path.join(data_root_path, "train_dataset.pkl")):
-        for i in range(args.n_property_samples_train * len(train_objects)):
-            path = os.path.join(train_objects_path, f"{i}.pkl")
-            with open(path, "rb") as handle:
-                object_data_list.append(pickle.load(handle))
-        # load up train objects make the set
-        # num_processes = multiprocessing.cpu_count()  # Use all available CPU cores
-        # results = pool.map(process_func, object_data_list)
-        with tqdm.tqdm(total=len(object_data_list), desc="Processing objects") as pbar:
-            results = []
-            for result in pool.imap_unordered(process_func, object_data_list):
-                results.append(result)
-                pbar.update()
+    # if not os.path.exists(os.path.join(data_root_path, "train_dataset.pkl")):
+    for i in range(args.n_property_samples_train * len(train_objects)):
+        path = os.path.join(train_objects_path, f"{i}.pkl")
+        with open(path, "rb") as handle:
+            object_data_list.append(pickle.load(handle))
+    # load up train objects make the set
+    # num_processes = multiprocessing.cpu_count()  # Use all available CPU cores
+    # results = pool.map(process_func, object_data_list)
+    with tqdm.tqdm(total=len(object_data_list), desc="Processing objects") as pbar:
+        results = []
+        for result in pool.imap_unordered(process_func, object_data_list):
+            results.append(result)
+            pbar.update()
 
-        path = os.path.join(data_root_path, "train_dataset.pkl")
-        with open(path, "wb") as handle:
-            pickle.dump(results, handle)
+    path = os.path.join(data_root_path, "train_dataset.pkl")
+    with open(path, "wb") as handle:
+        pickle.dump(results, handle)
     
     if not os.path.exists(os.path.join(data_root_path, "samegeo_test_dataset.pkl")):
         object_data_list = []
@@ -356,21 +267,21 @@ def main(args):
         with open(path, "wb") as handle:
             pickle.dump(results, handle)
 
-    # if not os.path.exists(os.path.join(data_root_path, "test_dataset.pkl")):
-    object_data_list = []
-    for i in range(args.n_property_samples_test * len(test_objects)):
-        path = os.path.join(test_objects_path, f"{i}.pkl")
-        with open(path, "rb") as handle:
-            object_data_list.append(pickle.load(handle))
+    if not os.path.exists(os.path.join(data_root_path, "test_dataset.pkl")):
+        object_data_list = []
+        for i in range(args.n_property_samples_test * len(test_objects)):
+            path = os.path.join(test_objects_path, f"{i}.pkl")
+            with open(path, "rb") as handle:
+                object_data_list.append(pickle.load(handle))
 
-    with tqdm.tqdm(total=len(object_data_list), desc="Processing objects") as pbar:
-        results = []
-        for result in pool.imap_unordered(process_func, object_data_list):
-            results.append(result)
-            pbar.update()
-    path = os.path.join(data_root_path, "test_dataset.pkl")
-    with open(path, "wb") as handle:
-        pickle.dump(results, handle)
+        with tqdm.tqdm(total=len(object_data_list), desc="Processing objects") as pbar:
+            results = []
+            for result in pool.imap_unordered(process_func, object_data_list):
+                results.append(result)
+                pbar.update()
+        path = os.path.join(data_root_path, "test_dataset.pkl")
+        with open(path, "wb") as handle:
+            pickle.dump(results, handle)
 
     pool.close()
     pool.join()
@@ -387,12 +298,12 @@ if __name__ == "__main__":
     # parser.add_arguments...
 
     # TODO: Figure out if all of these arguments are neccesary
-    parser.add_argument("--train-objects-fname", type=str, default='train_objects')
-    parser.add_argument("--test-objects-fname", type=str, default='test_objects')
+    parser.add_argument("--train-objects-fname", type=str, required=True)
+    parser.add_argument("--test-objects-fname", type=str, required=True)
     parser.add_argument("--data-root-name", type=str, required=True)
-    parser.add_argument("--n-property-samples-train", type=int, default=1)
-    parser.add_argument("--n-property-samples-test", type=int, default=1)
-    parser.add_argument("--n-pushes-per-object", type=int,default=3)
+    parser.add_argument("--n-property-samples-train", type=int, required=True)
+    parser.add_argument("--n-property-samples-test", type=int, required=True)
+    parser.add_argument("--n-pushes-per-object", type=int, required=True)
     parser.add_argument("--gui", action='store_true', default=False)
     # parser.add_argument('--n-points-per-object', type=int, required=True)
     # parser.add_argument('--n-fit-grasps', type=int, required=True)
